Add files via upload
The BandMatrix.cc class is the one created with supervisor this morning. It uses a "map" object to deal with band matrices so everything is very fast. All the rest of the files are the classes I tried to create in order to form the band matrix. They work but are really slow.